requirements.txt
# =============================================================================
# Core Framework
# =============================================================================
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# =============================================================================
# Database & ORM
# =============================================================================
sqlalchemy>=2.0.25
pymssql>=2.2.11
pyodbc>=5.0.1
alembic>=1.13.1
duckdb>=0.10.0  # For querying Gold Delta tables directly (replaces analytics.* SQL projection)

# =============================================================================
# Databricks & Delta Lake
# =============================================================================
databricks-sdk>=0.18.0
databricks-sql-connector>=3.0.0
pyspark==3.5.3  # Downgraded from 4.1.1 for hadoop-aws compatibility
delta-spark==3.2.1  # Compatible with PySpark 3.5.x
deltalake>=0.15.0  # delta-rs Python bindings

# =============================================================================
# Azure Services
# =============================================================================
azure-storage-blob>=12.19.0
azure-identity>=1.15.0
azure-monitor-opentelemetry>=1.2.0

# =============================================================================
# Data Processing
# =============================================================================
pandas>=2.1.4
numpy>=1.26.3
pyarrow>=14.0.2

# =============================================================================
# Authentication & Security
# =============================================================================
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6

# =============================================================================
# Caching & Session
# =============================================================================
redis>=5.0.1
hiredis>=2.3.2

# =============================================================================
# Testing
# =============================================================================
pytest>=7.4.3
pytest-asyncio>=0.23.3
pytest-cov>=4.1.0
httpx>=0.26.0

# =============================================================================
# Development Tools
# =============================================================================
python-dotenv>=1.0.0
black>=23.12.1
ruff>=0.1.11
mypy>=1.8.0

# =============================================================================
# Monitoring & Logging
# =============================================================================
python-json-logger>=2.0.7

# =============================================================================
# External Integrations
# =============================================================================
# sendgrid>=6.11.0
# twilio>=8.11.1

# =============================================================================
# S3 Compatible Storage (MinIO)
# =============================================================================
minio>=7.2.3
